fix(Card): :bug: fix card.content rotation animation
Fix that if two cards match, the card.content of the first face up
card also gets animated. This can be achieved by putting the ZStack
(and with it the card.content) of the CardView always on screen and
not only when the card.isFaceUp=true. Because in order that a view can
be animated, it has to be already on screen.